Added Data.Trie.Map, mostly copied from Set.
The reason for the line count being almost twice that of Set is
primarily due to the fact that what used to be just 'f' is often now
generalized to 'fWithKey', 'fBy', etc. The WithKey versions can't
usually be used to define the basic 'f' since they involve extra work to
keep track of the current key and avoiding that performance hit is
preferable. Hence there's a lot of similar-looking code.