A new class hierarchy for 'st', for Identity.
Since Identity often needs to specifically work with Bool, we can't use
the standard Alternative class which doesn't allow for constraints on
the contained type.

For the same reason, SetValue had to be split into three classes.
Boolable was fortunately a simple change and Unwrappable could be left
unchanged.

Unfortunately this means that we can't specify many of these as base
classes in the declaration of Trie, we have to uglify the types of the
functions instead.