New test properties: is[Proper]Sub(set|map)Of.
Changes to GetKey where made because I originally used
NoMonomorphismRestriction here: I ended up not needing it but it did
break two earlier properties, so I figured I'd fix them.

In order to not have to make isSubsetOf_t and co, I used (m :: TrieType)
in the body instead: this required a change to expandE in makeFunc.