Remove strict differences, same as normal ones.
They always need to evaluate the values to the point of seeing whether
they're Just or Nothing, in order to see whether a subtrie became null.

Exactly one case was lazy: non-Patricia tries where the differing keys
were at []. This is too obscure to be useful and thus non-Patricia
always seqs the values, for consistency.